refactor(icon-wrapper): fix requested code review changes
this PR improves the implementation by unifying the component used for both mobile and web since
it's the same, and also adds fill and stroke properties conditionally

refactor(icon-wrapper): improve wrapper applying suggested changes

this unify both web and native implementations, since the component is the same. also update fill
and stroke properties to be sent conditionally
