+++ packages/yoga/src/IconWrapper/native/index.js
@@ -1,3 +1,3 @@
-import IconWrapper from './IconWrapper';
+import IconWrapper from '../IconWrapper';
 
 export default IconWrapper;
